Remove tests
